+++ inst/doc/tidyxl.R
@@ -17,9 +17,8 @@ readxl::read_excel(titanic)
 
 ## ---- echo = TRUE--------------------------------------------------------
 library(tidyxl)
-x <- tidy_xlsx(titanic)$data$Sheet1
-# Specific sheets can be requested using `tidy_xlsx(file, sheet)`
-str(x)
+x <- xlsx_cells(titanic)
+dplyr::glimpse(x)
 
 ## ---- echo = TRUE--------------------------------------------------------
 x[x$data_type == "character", c("address", "character")]
@@ -27,7 +26,7 @@ x[x$row == 4, c("address", "character", "numeric")]
 
 ## ---- echo = TRUE--------------------------------------------------------
 # Bold
-formats <- tidy_xlsx(titanic)$formats
+formats <- xlsx_formats(titanic)
 formats$local$font$bold
 x[x$local_format_id %in% which(formats$local$font$bold),
   c("address", "character")]
@@ -42,14 +41,28 @@ x[x$local_format_id %in%
 formats$style$font$name["Normal"]
 head(x[x$style_format == "Normal", c("address", "character")])
 
+# In-cell formatting is available in the `character_formatted` column as a data
+# frame, one row per substring.
+examples <- system.file("/extdata/examples.xlsx", package = "tidyxl")
+xlsx_cells(examples)$character_formatted[77]
+
 ## ---- echo = TRUE--------------------------------------------------------
 x[!is.na(x$comment), c("address", "comment")]
 
 ## ---- echo = TRUE-----------------------------------------------------------------------------------------------------
 options(width = 120)
-y <- tidy_xlsx(system.file("/extdata/examples.xlsx", package = "tidyxl"),
-               "Sheet1")$data[[1]]
+y <- xlsx_cells(examples, "Sheet1")
 y[!is.na(y$formula),
-  c("address", "formula", "formula_type", "formula_ref", "formula_group",
+  c("address", "formula", "is_array", "formula_ref", "formula_group",
     "error", "logical", "numeric", "date", "character")]
+
+## ---- fig.width = 7, fig.height = 5-----------------------------------------------------------------------------------
+x <- xlex("MIN(3,MAX(2,A1))")
+x
+
+## ---------------------------------------------------------------------------------------------------------------------
+xlsx_names(examples)
+
+## ---------------------------------------------------------------------------------------------------------------------
+xlsx_validation(examples)