+++ src/border.cpp
@@ -1,15 +1,26 @@
 #include <Rcpp.h>
 #include "rapidxml.h"
 #include "border.h"
-#include "styles.h"
+#include "xlsxstyles.h"
 #include "stroke.h"
 
 using namespace Rcpp;
 
-border::border() {} // default constructor
+inline int bool_value(rapidxml::xml_node<>* node, const char* name) {
+  std::string value;
+  rapidxml::xml_attribute<>* attribute = node->first_attribute(name);
+  if (attribute == NULL) {
+    return(false);
+  }
+  value = attribute->value();
+  if (value == "0" || value == "false") {
+    return(false);
+  }
+  return(true);
+}
 
 border::border(rapidxml::xml_node<>* border,
-    styles* styles
+    xlsxstyles* styles
     ):
   left_(border->first_node("left"), styles),
   right_(border->first_node("right"), styles),
@@ -18,22 +29,12 @@ border::border(rapidxml::xml_node<>* border,
   top_(border->first_node("top"), styles),
   bottom_(border->first_node("bottom"), styles),
   diagonal_(border->first_node("diagonal"), styles),
-  vertical_(border->first_node("vertical"), styles),
-  horizontal_(border->first_node("horizontal"), styles)
+  vertical_(border->first_node("vertical"), styles),    // can't produce in Excel
+  horizontal_(border->first_node("horizontal"), styles) // can't produce in Excel
 {
-  diagonalDown_ = IntegerVector::create(NA_INTEGER);
-  rapidxml::xml_attribute<>* diagonalDown = border->first_attribute("diagonalDown");
-  if (diagonalDown != NULL)
-    diagonalDown_[0] = strtol(diagonalDown->value(), NULL, 10);
-
-  diagonalUp_   = IntegerVector::create(NA_INTEGER);
-  rapidxml::xml_attribute<>* diagonalUp = border->first_attribute("diagonalUp");
-  if (diagonalUp != NULL)
-    diagonalUp_[0] = strtol(diagonalUp->value(), NULL, 10);
+  diagonalDown_ = bool_value(border, "diagonalDown");
+  diagonalUp_ = bool_value(border, "diagonalUp");
 
   // I haven't been able to create an outline attribute in Excel 2016
-  outline_      = IntegerVector::create(NA_INTEGER);
-  rapidxml::xml_attribute<>* outline = border->first_attribute("outline");
-  if (outline != NULL)
-    outline_[0] = strtol(outline->value(), NULL, 10);
+  outline_ = bool_value(border, "outline");
 }