+++ src/border.h
@@ -1,7 +1,7 @@
 #ifndef BORDER_
 #define BORDER_
 
-class styles; // Forward declaration because fill is included in styles.
+class xlsxstyles; // Forward declaration because fill is included in xlsxstyles.
 
 #include <Rcpp.h>
 #include "rapidxml.h"
@@ -11,21 +11,20 @@ class border {
 
   public:
 
-    Rcpp::IntegerVector diagonalDown_;
-    Rcpp::IntegerVector diagonalUp_;
-    Rcpp::IntegerVector outline_;
-    stroke  left_;
-    stroke  right_;
-    stroke  start_; // gnumeric
-    stroke  end_;   // gnumeric
-    stroke  top_;
-    stroke  bottom_;
-    stroke  diagonal_;
-    stroke  vertical_;
-    stroke  horizontal_;
+    int    diagonalDown_;
+    int    diagonalUp_;
+    int    outline_;
+    stroke left_;
+    stroke right_;
+    stroke start_; // gnumeric
+    stroke end_;   // gnumeric
+    stroke top_;
+    stroke bottom_;
+    stroke diagonal_;
+    stroke vertical_;
+    stroke horizontal_;
 
-    border(); // Default constructor
-    border(rapidxml::xml_node<>* border, styles* styles);
+    border(rapidxml::xml_node<>* border, xlsxstyles* styles);
 };
 
 #endif