+++ src/color.h
@@ -1,8 +1,8 @@
 #ifndef COLOR_
 #define COLOR_
 
-class styles; // Forward declaration because color is included in font, which is
-              // included in styles.
+class xlsxstyles; // Forward declaration because color is included in font, which is
+              // included in xlsxstyles.
 
 #include <Rcpp.h>
 #include "rapidxml.h"
@@ -11,13 +11,13 @@ class color {
 
   public:
 
-    Rcpp::String        rgb_;
-    Rcpp::IntegerVector theme_;
-    Rcpp::IntegerVector indexed_;
-    Rcpp::NumericVector tint_;
+    Rcpp::String rgb_;
+    Rcpp::String theme_;
+    int          indexed_;
+    double       tint_;
 
     color(); // default constructor
-    color(rapidxml::xml_node<>* color, styles* styles);
+    color(rapidxml::xml_node<>* color, xlsxstyles* styles);
 };
 
 #endif