+++ src/font.h
@@ -1,7 +1,7 @@
 #ifndef FONT_
 #define FONT_
 
-class styles; // Forward declaration because font is included in styles.
+class xlsxstyles; // Forward declaration because font is included in xlsxstyles.
 
 #include <Rcpp.h>
 #include "rapidxml.h"
@@ -11,19 +11,18 @@ class font {
 
   public:
 
-    Rcpp::LogicalVector b_;         // bold
-    Rcpp::LogicalVector i_;         // italic
-    Rcpp::String        u_;         // underline (val attribute e.g. "none" or no attribute at all)
-    Rcpp::LogicalVector strike_;    // strikethrough
-    Rcpp::String        vertAlign_; // (val attribute)
-    Rcpp::IntegerVector size_;      // or sz for googlesheets
-    color               color_;
-    Rcpp::String        name_;
-    Rcpp::IntegerVector family_;
-    Rcpp::String        scheme_;
+    int          b_;         // bold
+    int          i_;         // italic
+    Rcpp::String u_;         // underline (val attribute e.g. "none" or no attribute at all)
+    int          strike_;    // strikethrough
+    Rcpp::String vertAlign_; // (val attribute)
+    double       size_;      // or sz for googlesheets
+    color        color_;
+    Rcpp::String name_;
+    int          family_;
+    Rcpp::String scheme_;
 
-    font(); // default constructor
-    font(rapidxml::xml_node<>* font, styles* styles);
+    font(rapidxml::xml_node<>* font, xlsxstyles* styles);
 };
 
 #endif