+++ src/stroke.cpp
@@ -1,16 +1,14 @@
 #include <Rcpp.h>
 #include "rapidxml.h"
 #include "stroke.h"
-#include "styles.h"
+#include "xlsxstyles.h"
 #include "color.h"
 
 using namespace Rcpp;
 
-stroke::stroke() {} // Default constructor
-
 stroke::stroke(
     rapidxml::xml_node<>* stroke,
-    styles* styles
+    xlsxstyles* styles
     ) {
   style_ = NA_STRING;