+++ src/xlsxcell.cpp
@@ -3,7 +3,8 @@
 #include "xlsxbook.h"
 #include "xlsxcell.h"
 #include "xlsxsheet.h"
-#include "utils.h"
+#include "string.h"
+#include "date.h"
 
 using namespace Rcpp;
 
@@ -13,9 +14,9 @@ xlsxcell::xlsxcell(
     xlsxbook& book,
     unsigned long long int& i
     ) {
-    parseAddress(cell, sheet, i);
+    parseAddress(cell, sheet, book, i);
     cacheValue  (cell, sheet, book, i); // Also caches format, as inextricable
-    cacheFormula(cell, sheet, i);
+    cacheFormula(cell, sheet, book, i);
 }
 
 // Based on hadley/readxl
@@ -25,34 +26,32 @@ xlsxcell::xlsxcell(
 void xlsxcell::parseAddress(
     rapidxml::xml_node<>* cell,
     xlsxsheet* sheet,
+    xlsxbook& book,
     unsigned long long int& i
     ) {
   rapidxml::xml_attribute<>* r = cell->first_attribute("r");
-  if (r == NULL)
-    stop("Invalid cell: lacks 'r' attribute");
-
   address_ = r->value(); // we need this std::string in a moment
-  sheet->address_[i] = address_;
+  book.address_[i] = address_;
 
+  col_ = 0;
+  row_ = 0;
   // Iterate though the A1-style address string character by character
-  int col = 0;
-  int row = 0;
   for(std::string::const_iterator iter = address_.begin();
       iter != address_.end(); ++iter) {
     if (*iter >= '0' && *iter <= '9') { // If it's a number
-      row = row * 10 + (*iter - '0'); // Then multiply existing row by 10 and add new number
+      row_ = row_ * 10 + (*iter - '0'); // Then multiply existing row by 10 and add new number
     } else if (*iter >= 'A' && *iter <= 'Z') { // If it's a character
-      col = 26 * col + (*iter - 'A' + 1); // Then do similarly with columns
+      col_ = 26 * col_ + (*iter - 'A' + 1); // Then do similarly with columns
     }
   }
-  sheet->col_[i] = col;
-  sheet->row_[i] = row;
+  book.col_[i] = col_;
+  book.row_[i] = row_;
 
   // Look up any comment using the address, and delete it if found
   std::map<std::string, std::string>& comments = sheet->comments_;
   std::map<std::string, std::string>::iterator it = comments.find(address_);
   if(it != comments.end()) {
-    SET_STRING_ELT(sheet->comment_, i, Rf_mkCharCE(it->second.c_str(), CE_UTF8));
+    SET_STRING_ELT(book.comment_, i, Rf_mkCharCE(it->second.c_str(), CE_UTF8));
     comments.erase(it);
   }
 }
@@ -68,9 +67,8 @@ void xlsxcell::cacheValue(
   std::string vvalue;
   if (v != NULL) {
     vvalue = v->value();
-    sheet->content_[i] = vvalue;
   } else {
-    sheet->content_[i] = NA_STRING;
+    book.is_blank_[i] = true;
   }
 
   // 't' for 'type' defines the meaning of 'v' for value
@@ -78,9 +76,6 @@ void xlsxcell::cacheValue(
   std::string tvalue;
   if (t != NULL) {
     tvalue = t->value();
-    sheet->type_[i] = tvalue;
-  } else {
-    sheet->type_[i] = NA_STRING;
   }
 
   // 's' for 'style' indexes into data structures of formatting
@@ -92,117 +87,123 @@ void xlsxcell::cacheValue(
   } else {
     svalue = 0;
   }
-  sheet->local_format_id_[i] = svalue + 1;
-  sheet->style_format_[i] = book.styles_.cellStyles_map_[book.styles_.cellXfs_[svalue].xfId_[0]];
+  book.local_format_id_[i] = svalue + 1;
+  book.style_format_[i] = book.styles_.cellStyles_map_[book.styles_.cellXfs_[svalue].xfId_];
 
   if (t != NULL && tvalue == "inlineStr") {
-    sheet->data_type_[i] = "character";
+    book.data_type_[i] = "character";
     rapidxml::xml_node<>* is = cell->first_node("is");
     if (is != NULL) { // Get the inline string if it's really there
       std::string inlineString;
       parseString(is, inlineString); // value is modified in place
-      SET_STRING_ELT(sheet->character_, i, Rf_mkCharCE(inlineString.c_str(), CE_UTF8));
+      SET_STRING_ELT(book.character_, i, Rf_mkCharCE(inlineString.c_str(), CE_UTF8));
     }
     return;
   } else if (v == NULL) {
     // Can't now be an inline string (tested above)
-    sheet->data_type_[i] = "blank";
+    book.data_type_[i] = "blank";
     return;
   } else if (t == NULL || tvalue == "n") {
-    if (book.styles_.cellXfs_[svalue].applyNumberFormat_[0] == 1) {
+    if (book.styles_.cellXfs_[svalue].applyNumberFormat_ == 1) {
       // local number format applies
-      if (book.styles_.isDate_[book.styles_.cellXfs_[svalue].numFmtId_[0]]) {
+      if (book.styles_.isDate_[book.styles_.cellXfs_[svalue].numFmtId_]) {
         // local number format is a date format
-        sheet->data_type_[i] = "date";
+        book.data_type_[i] = "date";
         double date = strtod(vvalue.c_str(), NULL);
-        sheet->date_[i] = checkDate(date, book.dateSystem_, book.dateOffset_,
-                                    ref(sheet->name_, address_));
+        book.date_[i] = checkDate(date, book.dateSystem_, book.dateOffset_,
+                                  "'" + sheet->name_ + "'!" + address_);
         return;
       } else {
-        sheet->data_type_[i] = "numeric";
-        sheet->numeric_[i] = strtod(vvalue.c_str(), NULL);
+        book.data_type_[i] = "numeric";
+        book.numeric_[i] = strtod(vvalue.c_str(), NULL);
       }
-    } else if (
+    } else if ( // no known case # nocov start
           book.styles_.isDate_[
             book.styles_.cellStyleXfs_[
-              book.styles_.cellXfs_[svalue].xfId_[0]
-            ].numFmtId_[0]
+              book.styles_.cellXfs_[svalue].xfId_
+            ].numFmtId_
           ]
         ) {
       // style number format is a date format
-      sheet->data_type_[i] = "date";
+      book.data_type_[i] = "date";
       double date = strtod(vvalue.c_str(), NULL);
-      sheet->date_[i] = checkDate(date, book.dateSystem_, book.dateOffset_,
-                                  ref(sheet->name_, address_));
+      book.date_[i] = checkDate(date, book.dateSystem_, book.dateOffset_,
+                                  "'" + sheet->name_ + "'!" + address_);
       return;
     } else {
-      sheet->data_type_[i] = "numeric";
-      sheet->numeric_[i] = strtod(vvalue.c_str(), NULL);
+      book.data_type_[i] = "numeric";
+      book.numeric_[i] = strtod(vvalue.c_str(), NULL); // # nocov end
     }
   } else if (tvalue == "s") {
     // the t attribute exists and its value is exactly "s", so v is an index
     // into the string table.
-    sheet->data_type_[i] = "character";
-    SET_STRING_ELT(sheet->character_, i, Rf_mkCharCE(book.strings_[strtol(vvalue.c_str(), NULL, 10)].c_str(), CE_UTF8));
+    book.data_type_[i] = "character";
+    SET_STRING_ELT(book.character_, i, Rf_mkCharCE(book.strings_[strtol(vvalue.c_str(), NULL, 10)].c_str(), CE_UTF8));
+    book.character_formatted_[i] = book.strings_formatted_[strtol(vvalue.c_str(), NULL, 10)];
     return;
   } else if (tvalue == "str") {
     // Formula, which could have evaluated to anything, so only a string is safe
-    sheet->data_type_[i] = "character";
-    sheet->character_[i] = vvalue;
+    book.data_type_[i] = "character";
+    book.character_[i] = vvalue;
     return;
   } else if (tvalue == "b"){
-    sheet->data_type_[i] = "logical";
-    sheet->logical_[i] = strtod(vvalue.c_str(), NULL);
+    book.data_type_[i] = "logical";
+    book.logical_[i] = strtod(vvalue.c_str(), NULL);
     return;
   } else if (tvalue == "e") {
-    sheet->data_type_[i] = "error";
-    sheet->error_[i] = vvalue;
+    book.data_type_[i] = "error";
+    book.error_[i] = vvalue;
     return;
-  } else if (tvalue == "d") {
+  } else if (tvalue == "d") { // # nocov start
     // Does excel use this date type? Regardless, don't have cross-platform
     // ISO8601 parser (yet) so need to return as text.
-    sheet->data_type_[i] = "date (ISO8601)";
-    return;
-  } else {
-    sheet->data_type_[i] = "unknown";
-    return;
+    book.data_type_[i] = "date (ISO8601)";
+    return; // # nocov end
+  } else { // no known case
+    book.data_type_[i] = "unknown"; // # nocov start
+    return; // # nocov end
   }
 }
 
 void xlsxcell::cacheFormula(
     rapidxml::xml_node<>* cell,
     xlsxsheet* sheet,
+    xlsxbook& book,
     unsigned long long int& i
     ) {
-  // TODO: Formulas are more complicated than this, because they're shared.
-  // p.1629 'shared' and 'si' attributes
-  // TODO: Array formulas use the ref attribute for their range, and t to
-  // state that they're 'array'.
   rapidxml::xml_node<>* f = cell->first_node("f");
+  std::string formula;
+  int si_number;
+  std::map<int, shared_formula>::iterator it;
   if (f != NULL) {
-    sheet->formula_[i] = f->value();
+    formula = f->value();
+    book.formula_[i] = formula;
     rapidxml::xml_attribute<>* f_t = f->first_attribute("t");
     if (f_t != NULL) {
-      sheet->formula_type_[i] = f_t->value();
-    } else {
-      sheet->formula_type_[i] = NA_STRING;
+      std::string ftvalue(f_t->value());
+      if (ftvalue == "array") {
+        book.is_array_[i] = true;
+      }
     }
+
     rapidxml::xml_attribute<>* ref = f->first_attribute("ref");
     if (ref != NULL) {
-      sheet->formula_ref_[i] = ref->value();
-    } else {
-      sheet->formula_ref_[i] = NA_STRING;
+      book.formula_ref_[i] = ref->value();
     }
+
+    // Formulas are sometimes defined once, and then 'shared' with a range
+    // p.1629 'shared' and 'si' attributes
     rapidxml::xml_attribute<>* si = f->first_attribute("si");
     if (si != NULL) {
-      sheet->formula_group_[i] = strtol(si->value(), NULL, 10);
-    } else {
-      sheet->formula_group_[i] = NA_INTEGER;
+      si_number = strtol(si->value(), NULL, 10);
+      book.formula_group_[i] = si_number;
+      if (formula.length() == 0) { // inherits definition
+        it = sheet->shared_formulas_.find(si_number);
+        book.formula_[i] = it->second.offset(row_, col_);
+      } else { // defines shared formula
+        shared_formula new_shared_formula(formula, row_, col_);
+        sheet->shared_formulas_.insert({si_number, new_shared_formula});
+      }
     }
-  } else {
-    sheet->formula_[i] = NA_STRING;
-    sheet->formula_type_[i] = NA_STRING;
-    sheet->formula_ref_[i] = NA_STRING;
-    sheet->formula_group_[i] = NA_INTEGER;
   }
 }