+++ src/xlsxcell.h
@@ -9,6 +9,8 @@
 class xlsxcell {
 
   std::string address_;
+  int col_;
+  int row_;
 
   public:
 
@@ -19,14 +21,27 @@ class xlsxcell {
         unsigned long long int& i   // the index of the cell
         );
 
-    void parseAddress( rapidxml::xml_node<>* cell, xlsxsheet* sheet, unsigned long long int& i);
-    void cacheValue( rapidxml::xml_node<>* cell, xlsxsheet* sheet, xlsxbook& book, unsigned long long int& i);
+    void parseAddress(
+        rapidxml::xml_node<>* cell,
+        xlsxsheet* sheet,
+        xlsxbook& book,
+        unsigned long long int& i
+        );
 
-    // Not the whole format, but only the index of the local and style
-    // formats (cellXfs and cellStyleXfs)
-    void cacheFormat( rapidxml::xml_node<>* cell, xlsxsheet* sheet, xlsxbook& book, unsigned long long int& i);
+    void cacheValue(
+        rapidxml::xml_node<>* cell,
+        xlsxsheet* sheet,
+        xlsxbook& book,
+        unsigned long long int& i
+        );
 
-    void cacheFormula( rapidxml::xml_node<>* cell, xlsxsheet* sheet, unsigned long long int& i);
+    void cacheFormula(
+        rapidxml::xml_node<>* cell,
+        xlsxsheet* sheet,
+        xlsxbook& book,
+        unsigned long long int& i
+        );
+
 };
 
 #endif