+++ src/xlsxstyles.h
@@ -1,5 +1,5 @@
-#ifndef STYLES_
-#define STYLES_
+#ifndef XLSXSTYLES_
+#define XLSXSTYLES_
 
 #include <Rcpp.h>
 #include "rapidxml.h"
@@ -10,17 +10,25 @@
 
 struct colors {
   Rcpp::CharacterVector rgb;
-  Rcpp::IntegerVector   theme;
+  Rcpp::CharacterVector theme;
   Rcpp::IntegerVector   indexed;
   Rcpp::NumericVector   tint;
+  colors() {}
+  colors(int n):
+    rgb     (n, NA_STRING),
+    theme   (n, NA_STRING),
+    indexed (n, NA_INTEGER),
+    tint    (n, NA_REAL)
+  {}
 };
 
-class styles {
+class xlsxstyles {
 
   public:
 
-    Rcpp::CharacterVector theme_;   // rgb equivalent of theme no.
-    Rcpp::CharacterVector indexed_;   // rgb equivalent of index no.
+    Rcpp::CharacterVector theme_name_; // name of theme, e.g. "accent1"
+    Rcpp::CharacterVector theme_;      // rgb equivalent of theme no.
+    Rcpp::CharacterVector indexed_;    // rgb equivalent of index no.
 
     std::vector<xf> cellXfs_;
 
@@ -35,32 +43,30 @@ class styles {
     std::vector<fill> fills_;
     std::vector<border> borders_;
 
-    xf style_formats_; // built up by applyFormats() from xf definitions
-    xf local_formats_; // built up by applyFormats() from xf definitions
+    std::vector<xf> style_formats_; // built up by applyFormats() from xf definitions
+    std::vector<xf> local_formats_; // built up by applyFormats() from xf definitions
 
     Rcpp::List style_; // inside-out List version of style_formats_
     Rcpp::List local_; // inside-out List version of local_formats_
 
-    styles(const std::string& path);
+    xlsxstyles(const std::string& path);
 
     void cacheThemeRgb(const std::string& path);
     void cacheIndexedRgb();
 
     void cacheCellXfs(rapidxml::xml_node<>* styleSheet);
     void cacheCellStyleXfs(rapidxml::xml_node<>* styleSheet); // also cellStyles, if available
-
     void cacheNumFmts(rapidxml::xml_node<>* styleSheet);
-    bool isDateFormat(std::string formatCode);
-
     void cacheFonts(rapidxml::xml_node<>* styleSheet);
     void cacheFills(rapidxml::xml_node<>* styleSheet);
     void cacheBorders(rapidxml::xml_node<>* styleSheet);
 
-    void clone_color(color& from, colors& to); // Append values in one color to vectors in another
+    // Insert values of one color object into vectors in another
+    void clone_color(color& from, colors& to, int& i);
     Rcpp::List list_color(colors& original, bool is_style); // Convert struct to list
 
     void applyFormats(); // Build each style on top of the normal style
-    Rcpp::List zipFormats(xf styles, bool is_style); // Turn the formats inside-out to return to R
+    Rcpp::List zipFormats(std::vector<xf> styles, bool is_style); // Turn the formats inside-out to return to R
 
 };